Update main.py
Setting my onboard LED to help me ensure that the Raspberry PICO is running as expected. prior to that my OnboardLED function I have added same variables to help me control the values and sped of my external LEDs. 

This program is set to run automatically when it receives power and will flow with the following light effects in a timely mannered fashion.